feat(auth): refactor unlinkAccountAction to use service layer for business logic
